In test script, exit with error if distribution not present
This pertains to the case where `tangled test` is run without specifying
where the tests are. In this case, we attempt to guess where the tests
are based on the name of the distribution in the current directory,
which requires `pip install -e .` (or similar) to be run first.

Previously, this situation would cause an unhelpful error due to a bug
where the dist would be set to `True` if no dist was found in the
current directory. Now, we exit with a hint.