Allow cached properties specify their dependencies
This allows a cached property to be reset automatically when any of its
dependencies are set or deleted.

There's one backward-incompatible change. Previously, deleting a
property that had not been set (via access or directly) would raise an
`AttributeError`, but now it won't. This is because it's always safe to
access a cached property (unlike a normal instance attribute where if
it's deleted, attempting to access it will fail).