Make it easy to reset cached properties dependent on regular attributes
Previously, if a cached property depended on a regular attribute, when
the regular attribute was updated, there was no easy way to reset the
cached property. Now, it's possible to do so with simple `__setattr__`
and `__delattr__` overrides that use the new
`cached_property.reset_dependents_of()` class method.

For example:

    class C:

        def __init__(self):
            self.x = 1

        @cached_property('x')
        def prop(self):
            return self.x

        def __setattr__(self, name, value):
            super().__setattr__(name, value)
            cached_property.reset_dependents_of(self, name)

        def __delattr__(self, name):
            super().__delattr__(name)
            cached_property.reset_dependents_of(self, name)