Simplify/improve internals of cached_property
The main change is that a map of dependencies is now stored on classes
that define cached properties. This simplifies resetting dependents
since we don't have to go through _all_ of a class's attributes to find
the cached properties. The logic for resetting dependents is now a bit
easier to understand and should be faster too.