Return result of command's run method from main()
If a command's `run` method returns nothing, that's equivalent to
returning `0`.

When running a tangled script via the command line, the return value of
the command's `run` method will be used as the script's exit code.

Commands that use `ACommand.exit(code)` should instead return the exit
code from `run`.