Revamp/improve/simplify settings parsing
Settings values in config files must now be encoded as JSON.

For special cases, settings keys can include a string-to-type converter
using the following syntax (similar to Python function/parameter
annotation syntax):

    name:type = "value"

This change allows all of the complex converter logic to be jettisoned.
Using JSON supports most use cases, and the ability to specify other
converters should cover the rest.

The `as_bool()` converter is still useful, so it was moved to the `util`
module.